Refactor module's name
