+++ docs/source/quintoandar_hierarchical_conf.rst
@@ -4,7 +4,7 @@ quintoandar\_hierarchical\_conf package
 Module contents
 ---------------
 
-.. automodule:: quintoandar_hierarchical_conf
+.. automodule:: hierarchical_conf
    :members:
    :undoc-members:
    :show-inheritance:
